rules modification and migraiton number
