+++ apps/backend-ops/src/firestoreMigrations/index.ts
@@ -13,7 +13,7 @@ import * as v0012 from './0012';
 import * as v0013 from './0013';
 import * as v0014 from './0014';
 import * as v0015 from './0015';
-import * as v0016 from './0016';
+import * as v0017 from './0017';
 
 import { Firestore } from '../admin';
 
@@ -41,7 +41,7 @@ export const MIGRATIONS = {
   13: v0013,
   14: v0014,
   15: v0015,
-  16: v0016,
+  17: v0017,
 };
 
 export const VERSIONS_NUMBERS = Object.keys(MIGRATIONS).map(s => parseInt(s, 10));
