+++ firestore.rules
@@ -183,8 +183,8 @@ service cloud.firestore {
       allow write: if false;
     }
 
-    /// DISTRIBUTION DEALS RULES ///
-    match /{path=**}/distributionDeals/{dealId} {
+    /// DISTRIBUTION RIGHTS RULES ///
+    match /{path=**}/distributionRights/{rightId} {
     	allow write: if true; // @TODO #1388 Only if contract.status  === 'unknown' ?
       allow read: if true; // @TODO #1388 Only if licensee.orgId == currentUser.orgId || if licensor.orgId == currentUser.orgId
     }
